services/backend/requirements.txt
fastapi==0.111.0
python-dotenv==1.0.1
SQLAlchemy==2.0.30
uvicorn==0.30.1
psycopg2-binary==2.9.9
pyjwt==2.9.0
passlib==1.7.4
pandas==2.2.3
openpyxl==3.1.2
reportlab==4.2.5
xlsxwriter==3.2.0